rename file too
